Update package dependencies and modify animations
The commit replaces the 'animejs' animation library with 'framer-motion' for better performance and control. Additionally, it changes the package name from 'startup-nextjs-template' to 'sec-ieee-hkn'. These changes are reflected in the package.json and package-lock.json files and the usage of the new animation library is implemented in the Hero component.